requirements.txt
pyTelegramBotAPI==4.5.0
oauth2client==4.1.3
pickle-mixin==1.0.2
google-api-core==2.7.2
google-api-python-client==2.45.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
googleform==0.0.1a2
schedule==1.1.0
thread6==0.2.0
selenium=4.2.0
